Refactor: Transform repository into a pip-installable Python package
This commit refactors the existing TOPSIS implementation into a proper Python package that can be installed using pip.

The main changes include:
- Restructuring the project to follow a standard Python package layout with a `topsis` source directory.
- Adding a `pyproject.toml` file to define package metadata and dependencies (`pandas`, `numpy`, `openpyxl`).
- Updating all imports to use relative imports within the package.
- Removing example usage code from the main library and placing it in a new `examples` directory.
- Updating the `README.md` with installation and usage instructions.
